+++ Leomard/Models/Lemmy/Forms/ModTools/DistinguishComment.swift
@@ -1,0 +1,13 @@
+//
+//  DistinguishComment.swift
+//  Leomard
+//
+//  Created by Konrad Figura on 27/07/2023.
+//
+
+import Foundation
+
+struct DistinguishComment: Codable {
+    let commentId: Int
+    let distinguished: Bool
+}